Cleanup and consistency changes
If you are interested in these changes, I am using your plugin with Mura 6.2 and found that the protocol and http_host are not needed when you call the getImageURL() Mura method. It returns the entire URL for you already.  I don't know if this is different behavior from previous versions of Mura or not.

While I was modifying this file I also noticed some inconsistencies on when new lines #NL# were being added to the end of the <meta> tags. They were not being added in all cases.

I also noticed that it was calling the getValue() method in some cases but specific get{attribute_name} in others. I changed the getValue() calls to use the get{attribute_name} methods instead.

I changed the $.content() method calls for the summary and body content to use the $.setDynamicContent() method instead. That way any [Mura] tags embedded in the content will be processed. Don't know if you intentionally were not using this method for some reason?

Just FYI if you want these changes or not.